+++ lib/extensions/context_ext.dart
@@ -16,4 +16,24 @@ extension ContextExt on BuildContext {
   /// Example: `context.isDarkMode`
   bool get isDarkMode =>
       MediaQuery.of(this).platformBrightness == Brightness.dark;
+
+  /// Get if keyboard is visible
+  ///
+  /// Example: `context.isKeyboardVisible`
+  bool get isKeyboardVisible => MediaQuery.of(this).viewInsets.bottom != 0;
+
+  /// Get the height of the keyboard
+  ///
+  /// Example: `context.keyboardHeight`
+  double get keyboardHeight => MediaQuery.of(this).viewInsets.bottom;
+
+  /// Check if the device orientation is landscape
+  ///
+  /// Example: `context.isLandscape`
+  bool get isLandscape => orientation == Orientation.landscape;
+
+  /// Check if the device orientation is portrait
+  ///
+  /// Example: `context.isPortrait`
+  bool get isPortrait => orientation == Orientation.portrait;
 }